misc changes

+++ src/server/game/Entities/Player/Player.cpp
@@ -3112,7 +3112,7 @@ bool Player::CanInteractWithQuestGiver(Object* questGiver)
     return false;
 }
 
-Creature* Player::GetNPCIfCanInteractWith(uint64 guid, uint32 npcflagmask)
+Creature* Player::GetNPCIfCanInteractWith(uint64 guid, uint64 npcflagmask)
 {
     // unit checks
     if (!guid)
